requirements.txt
adsputils==1.2.8
requests==2.25.1
psycopg2-binary==2.8.6
grapi==0.1.4
google-api-python-client==1.12.8
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
oauth2client==4.1.3
